requirements.txt
python-dotenv==1.0.0
schedule==1.2.0
PyYAML==6.0.1
runwayml==3.9.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
requests==2.31.0
gigachat==0.1.9 